+++ src/Field/Data/AssetLinkData.php
@@ -6,5 +6,7 @@ final class AssetLinkData
 {
 	public function __construct (
 		public readonly string $url,
+		public readonly ?int $width,
+		public readonly ?int $height,
 	) {}
 }