+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 2.3.0
 =====
 
-* (feature) Extract image dimensions from Storyblok URLs in `AssetData`.
+* (feature) Extract image dimensions from Storyblok URLs into `AssetData` and `AssetLinkData`.
 
 
 2.2.0